requirements.txt
Flask>=3.1,<4.0
Jinja2>=3.1,<4.0
pydantic>=2.12,<3.0
google-api-python-client>=2.169,<3.0
google-auth>=2.40,<3.0
google-auth-oauthlib>=1.2,<2.0
pypdf>=5.4,<6.0
playwright>=1.58,<2.0
